browser: Use a separate user-data-dir
Use a separate user-data-dir in order to keep chrome from launching the
new window as a child of an existing process, if one exists. This is
necessary to track the pid of the browser.

Previously the app would assume the browser had closed and report
incorrect status to the 2nd screen.